Eliminate the remaining GCC warnings I encountered under FreeBSD/amd64 (but which are *not* FreeBSD specific): the printf(3) format specifier "%llu" expects an "unsigned long long int" but an "uint64_t" argument can be just an "unsigned long int" on a true 64-bit platform. On a 32-bit platform "unsigned long int" is usually just 32-bit while "unsigned long long int" has to be used for the 64-bit type. But on a true 64-bit platform the "unsigned long int" can be already 64-bit (and be the same as "unsigned long long int"). Hence, one has to cast the "uint64_t" to the "unsigned long long int" to make the strict compilers happy, although the effective type sizes do not really mismatch...

+++ rpmdb/hdrNVR.c
@@ -91,7 +91,7 @@ int headerMacrosLoad(Header h)
 	
 	if (val) {
 	    if (val == numbuf)
-		sprintf(numbuf, "%llu", ival);
+		sprintf(numbuf, "%llu", (unsigned long long)ival);
 	    addMacro(NULL, tagm->macroname, NULL, val, -1);
 	}
 	he->p.ptr = _free(he->p.ptr);
